get to work properly
